+++ common/constants/project.py
@@ -1,5 +1,5 @@
-import os.path
+from pathlib import Path
 
 
-PROJECT_ROOT = os.path.join("/", *os.path.abspath(__file__).split(os.sep)[:-3])
-GLOBAL_CONFIG_FILE = os.path.join(PROJECT_ROOT, "pyproject.toml")
+PROJECT_ROOT = Path(*Path(__file__).resolve().parts[:-3])
+GLOBAL_CONFIG_FILE = PROJECT_ROOT / "pyproject.toml"