+++ common/exceptions/profile.py
@@ -1,2 +1,8 @@
-class MissingProfileException(Exception):
+from common.exceptions import NotFoundError
+
+
+class MissingProfileError(NotFoundError):
+    pass
+
+class MissingElementError(NotFoundError):
     pass